server/requirements.txt
asgiref==3.7.2
autopep8==2.0.2
Django==3.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
pycodestyle==2.10.0
pytz==2023.3
sqlparse==0.4.4
